# mock_core/openapi.yaml
openapi: 3.0.3
info:
  title: TransNova Logistics Core Mock API
  version: 1.0.0
  description: Shared deterministic in-memory system-of-record facade for the multi-agent logistics demo.
paths:
...
  /shipments/{tracking_number}:
    get:
      operationId: getShipment
      parameters:
        - {in: path, name: tracking_number, required: true, schema: {type: string}}
      responses:
        '200': {description: Shipment}
  /shipments/{tracking_number}/events:
    get:
      operationId: getTrackingEvents
      parameters:
        - {in: path, name: tracking_number, required: true, schema: {type: string}}
      responses:
        '200': {description: Tracking events}
  /shipments/{tracking_number}/sla:
    get:
      operationId: getShipmentSla
      parameters:
        - {in: path, name: tracking_number, required: true, schema: {type: string}}
      responses:
        '200': {description: SLA projection}
  /shipments/{tracking_number}/recovery-options:
    get:
      operationId: getRecoveryOptions
      parameters:
        - {in: path, name: tracking_number, required: true, schema: {type: string}}
      responses:
        '200': {description: Recovery options}
...